app/services/speech_recognition/requirements.txt
# Core Audio Processing
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
soundfile>=0.12.0
pandas>=1.3.0
numba>=0.56.0

# Audio I/O & Preprocessing
ffmpeg-python>=0.2.0
pydub>=0.25.1

# ASR Engines
openai-whisper>=20231117
faster-whisper>=0.9.0
ctranslate2>=3.24.0

# Speaker Diarization & Embeddings
scikit-learn>=1.0.0
# Pin PyTorch family for SpeechBrain compatibility; install CUDA wheels from PyTorch index if needed
torch>=2.0.0,<2.2.0
torchaudio>=2.0.0,<2.2.0
speechbrain>=0.5.16,<0.6.0
resemblyzer>=0.1.1.dev0

# Utilities / Monitoring
psutil>=5.8.0
GPUtil>=1.4.0

# Progress bars
tqdm>=4.62.0

# Optional Web API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
